fix: 修复关闭 api 访问失败的问题

+++ packages/service/src/modules/projects/projects.controller.ts
@@ -111,7 +111,7 @@ export class ProjectsController {
         await this.projectsService.createApiAccessPath(`/${project.apiAccessPath}`)
       }
 
-      if (enableApiAccess === false) {
+      if (enableApiAccess === false && project.apiAccessPath) {
         await this.projectsService.deleteApiAccessPath(`/${project.apiAccessPath}`)
       }
     }
